+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -42,6 +42,8 @@ const { DirForm } = require('./directives/DirForm');
 const { DirCaptureUserReply } = require('./directives/DirCaptureUserReply');
 const { DirMake } = require('./directives/DirMake');
 const { DirReplaceBotV2 } = require('./directives/DirReplaceBotV2');
+const { DirHubspot } = require('./directives/DirHubspot');
+const { DirCustomerio } = require('./directives/DirCustomerio');
 const { DirAskGPTV2 } = require('./directives/DirAskGPTV2');
 
 class DirectivesChatbotPlug {
@@ -608,7 +610,6 @@ class DirectivesChatbotPlug {
       });
     }
     else if (directive_name === Directives.ASK_GPT_V2) {
-
       new DirAskGPTV2(context).execute(directive, async (stop) => {;
         if (context.log) { console.log("AskGPTV2 stop?", stop);}
         if (stop == true) {
@@ -664,6 +665,24 @@ class DirectivesChatbotPlug {
           let next_dir = await this.nextDirective(this.directives);
           this.process(next_dir);
         }
+      })
+    }
+    else if (directive_name === Directives.HUBSPOT) {
+      new DirHubspot(context).execute(directive, async (stop) => {
+        if (context.log) { console.log("Hubspot stop?", stop);}
+        if (stop == true) {
+          if (context.log) { console.log("Stopping Actions on:", JSON.stringify(directive));}
+          this.theend();
+        } else {
+          let next_dir = await this.nextDirective(this.directives);
+          this.process(next_dir);
+        }
+      })
+    }
+    else if (directive_name === Directives.CUSTOMERIO) {
+      new DirCustomerio(context).execute(directive, async () => {
+        let next_dir = await this.nextDirective(this.directives);
+        this.process(next_dir);
       })
     }
     else {